Handle HTTP error status codes
Improve the error handling in the HTTP client slightly to also inspect
the HTTP error codes. For these APIs we only assume `200 OK` is what we
need.

This was found by trying to bump the number of hits from the places
query beyond the accepted API limit. This change also ups the
`hitsPerPage` to improve the likelihood of successful matches.