Fetch remaining address pages in parallel
Instead of fetching the remainder of pages sequentially, this change
computes the expected remainder of requests to make and initiates HTTP
requests for all remaining pages of data.

Note: Since we cannot know the size of the query result set up front we
always need to make the initial first request to get the `TotalHits`
count.

Once all requests have been fired off, it allocates space for the
remainder of the addresses, then proceeds to block on the first
requests.

Once the first request is finished the addresses in the response is
added to the result set and the loop proceeds to consume the results
from the second request, and so on.

There's not really much need for a `select` and process requests in the
order they complete because w need to wait for all requests regardless.
Whichever request is the slowest will block the reading, but then all
remaining requests to process will be ready to be processed.